refactor: 토큰 검증 로직 책임을 `JwtTokenProvider` 쪽으로 통합
