+++ backend/src/main/java/com/woowacourse/naepyeon/controller/auth/LoginInterceptor.java
@@ -1,10 +1,7 @@
-package com.woowacourse.naepyeon.controller;
+package com.woowacourse.naepyeon.controller.auth;
 
-import com.woowacourse.naepyeon.exception.TokenInvalidExpiredException;
-import com.woowacourse.naepyeon.exception.TokenInvalidFormException;
 import com.woowacourse.naepyeon.support.AuthorizationExtractor;
 import com.woowacourse.naepyeon.support.JwtTokenProvider;
-import java.util.Objects;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
@@ -24,20 +21,7 @@ public class LoginInterceptor implements HandlerInterceptor {
             return true;
         }
         final String token = AuthorizationExtractor.extract(request);
-        validateNotToken(token);
-        validateAvailableToken(token);
+        jwtTokenProvider.validateAbleToken(token);
         return true;
-    }
-
-    private void validateNotToken(final String token) {
-        if (Objects.isNull(token)) {
-            throw new TokenInvalidFormException();
-        }
-    }
-
-    private void validateAvailableToken(final String token) {
-        if (jwtTokenProvider.checkExpiredToken(token)) {
-            throw new TokenInvalidExpiredException();
-        }
     }
 }
